requirements.txt
Bio==1.6.0
biopython==1.81
firebase_admin==6.2.0
mysql_connector_repackaged==0.3.1
networkx==3.1
nltk==3.8.1
numpy==1.26.2
pandas==2.1.3
Requests==2.31.0
scikit_learn==1.3.1
streamlit==1.28.2
streamlit_lottie==0.0.5
streamlit_option_menu==0.3.6
torch==2.1.0
transformers==4.34.0
